requirements.txt
azure-iot-device==2.11.0
python-dotenv==1.0.0
